BugID:17646749:[utils_time]modify some var to uint64_t to prevent time overflow

+++ middleware/linkkit/sdk-c/src/infra/utils/misc/utils_timer.h
@@ -11,22 +11,22 @@
 #include "iot_import.h"
 
 typedef struct {
-    uint32_t time;
+    uint64_t time;
 } iotx_time_t;
 
 
 void iotx_time_start(iotx_time_t *timer);
 
-uint32_t utils_time_spend(iotx_time_t *start);
+uint64_t utils_time_spend(iotx_time_t *start);
 
-uint32_t iotx_time_left(iotx_time_t *end);
+uint64_t iotx_time_left(iotx_time_t *end);
 
-uint32_t utils_time_is_expired(iotx_time_t *timer);
+int utils_time_is_expired(iotx_time_t *timer);
 
 void iotx_time_init(iotx_time_t *timer);
 
-void utils_time_countdown_ms(iotx_time_t *timer, uint32_t millisecond);
+void utils_time_countdown_ms(iotx_time_t *timer, uint64_t millisecond);
 
-uint32_t utils_time_get_ms(void);
+uint64_t utils_time_get_ms(void);
 
 #endif /* _IOTX_COMMON_TIMER_H_ */
